+++ src/auth/index.ts
@@ -1,5 +1,6 @@
 /** @packageDocumentation */
 export type { AccountInfo, UploadUrlEntry } from './account-info.ts'
 export { InMemoryAccountInfo } from './in-memory.ts'
+export type { RealmName } from './realms.ts'
 export { getRealmUrl, REALM_URLS } from './realms.ts'
 export { UploadUrlPool } from './upload-url-pool.ts'