+++ src/index.ts
@@ -28,6 +28,7 @@ export {
   AccessDeniedError,
   B2Error,
   B2InsufficientCapabilityError,
+  B2RealmConfigurationError,
   B2SsrfError,
   BadAuthTokenError,
   BadBucketIdError,